+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import router from './components/Router';
+import router from './router/Router';
 import { RouterProvider } from 'react-router-dom';
 import './index.css';
